requirements.txt
networkx==2.5.1
numpy==1.18.5
pandas==1.1.5
matplotlib
torch==1.9.0
keras==2.4.3
gym==0.18.3
geopy==2.2.0
PyYAML==5.4.1
pudb
jupyter
seaborn
tqdm==4.36.1
stable-baselines3==1.1.0
tensorboard==2.6.0